+++ vue-components/src/components/MenuItem.ts
@@ -1,4 +1,5 @@
 export type MenuItem = {
 	label: string;
 	description: string;
+	tag?: string; // for internal use
 };