+++ vue-components/tests/e2e/specs/Button.test.js
@@ -1,9 +1,13 @@
-module.exports = {
-	'Button is visible': ( client ) => {
+describe( 'Button', function () {
+	beforeEach( function ( client ) {
 		client
-			.init()
-			.openComponentStory( 'button' )
+			.pause( 500 )
+			.init( client.launch_url + '/iframe.html?id=button--normal&viewMode=story' );
+	} );
+
+	it( 'is visible', function ( client ) {
+		client
 			.waitForElementPresent( '.wikit-Button' )
 			.assert.visible( '.wikit-Button' );
-	},
-};
+	} );
+} );